pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rabbitai"
version = "0.1.0"
description = "AI-powered CLI troubleshooting assistant"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "RabbitAI Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "langchain>=0.3.0,<1.0.0",
    "langgraph>=0.2.0,<1.0.0",
    "langchain-google-genai>=2.0.0,<3.0.0",
    "langchain-community>=0.3.0",
    "prompt_toolkit>=3.0.52",
    "rich>=14.2.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
rabbit = "rabbitai.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "black>=25.9.0",
    "flake8>=7.3.0",
    "mypy>=1.18.2",
]

[tool.setuptools]
packages = ["rabbitai", "rabbitai.llm", "rabbitai.tools", "rabbitai.context", "rabbitai.command_config", "rabbitai.agents"]
